+++ NEWS.md
@@ -2,6 +2,10 @@
 
 ## Features
 
+* New `odbcConnectionColumns()` function to describe the column types. This
+  data is used when binding columns, which allows drivers which do not support
+  the `SQLDescribeParam()` ODBC function, such as freeTDS. (#313, @detule)
+
 * Added a Teradata `odbcDataType()` to support writing logical data to Teradata servers (#240, @blarj09).
 
 * Subseconds are now retained when inserting POSIXct objects (#130, #208)